refactor(jest): remove version qualifiers in v27 facade impl (#4950)
remove the import aliases used in `jest-facade.ts`. the aliases were
renaming the imports to be '27-specific'. however, we're already in a
'jest-27' directory and working on the `Jest27Stencil` class (which is
kept with the '27' to help differentiate this class further). having
these extra import aliases made the instructions for adding a new
version of jest more onerous than it needed to be