refactor(worker): document worker code, improve types (#5034)
This documents a bunch of code related to our node.js worker setup
(which was previously not super well-documented) and also makes a few
improvements to some relevant types.

This was motivated by #5010, which entailed modifying some of this code.
It wasn't entirely apparent how the code in `src/sys/node/worker.ts`,
well, _worked_ and what it was supposed to do. Now we know!