chore(snc): fix some boolean-related snc errors
This fixes a bunch of errors caused by passing `boolean | undefined`
where a property or parameter is actually typed `boolean`. This isn't
likely to actually cause too many issues at runtime since `undefined` is
falsy, but it does cause `strictNullChecks` errors that we'd like to get
rid of.

To fix the issues we can just do a `!!` coercion of the falsy
`undefined` value to `false`.