chore(compiler): fixup strictNullChecks errors in build-stats.ts (#3809)
This fixes up strictNullChecks violations in one file
(`src/compiler/build/build-stats.ts`) and then deals with some changes
across the codebase that follow from that.

In particular, I converted that file to use the `ValidatedConfig` type
and then made the fields on that interface accessed within the file
required, fixing some strict null checks violations in the file. This
then required changing from `Config` to `ValidatedConfig` in a few most
call sites, and changing a few things to support that about how the
config validation works.