fix(testing): make Puppeteer an optional dependency
Currently it seems required to have Puppeteer installed to build a Stencil component. It seems that the compiler imports parts of the testing module to compile a project which causes "/testing/puppeteer/puppeteer-declarations.d.ts" to be loaded.

To solve the type dependency the easiest is to add a Rollup plugin that simply adds a "@ts-ignore" before the import. If Puppeteer is not installed, it will not throw any errors, if it is installed, all types are propagated correctly.

STENCIL-881
fixes #4526