chore(snc): resolve errors for optimize-css tests
resolve `strictNullCheck` violations for `optimize-css#optimizeCss`
unit tests. the `fileName` argument doesn't appear to be used by
`optimizeCss` in any meaningful way that (other than an additional path
normalization call).

`optimizeCss` is a part of stencil's public api. as a result, we cannot
remove this otherwise unused function parameter without it constituting
a breaking change. add a TODO for now, to be completed in the next
major version of stencil.

add a type assertion to invalid styling input. this allows us to test
invalid css being passed to the optimizatiton function, without
widening the type signature of `optimizeCss`