chore(compiler): make dependency fields non-optional
make the following fields non-optional on `ComponentCompilerMeta`:
- dependents
- dependencies
- directDependents
- directDependencies

these fields were previously marked as optional as they are not
set when the metadata for a component is created, as we need to
generate the metadata for all components in order to set these fields.

with this constraint in mind, we stub the initial fields with an empty
array.