feat(compiler): Stencil decorator import aliasing (#5161)
* create map of original imports to their aliased import

this commit creates a map of original import names to aliased import names. This will be used to get the aliased name of a decorator when parsing Stencil decorators to static component metadata

* change `importAliasMap` to class

This commit creates an `ImportAliasMap` class that extends the `Map` class. This makes it easy to instantiate, reference the underlying type, and clearly override the default getter

* pass aliased name to decorator parsers

The commit passes the aliased import name from the `ImportAliasMap` to each respective decorator transformer

Fixes: #3137

STENCIL-263

* add e2e tests for some aliased imports

* check namedBindings exist

* add rest of decorators to test