test(ci): smoke test that Stencil doesn't error when installed alone
This adds a smoke test that checks that a project with only
`@stencil/core` installed will build a simple component without errors.

When working on #5165 it became clear we should have a regression test
in place to ensure we know right away if a change makes a 'bare' stencil
project (i.e. with only `@stencil/core` installed) is going to error
out.