chore(snc): resolve hmr logging-related errors (#5185)
update `build-finish#logHmr` to take a reference to the
`HotModuleReplacement` object from the build context, as we check for
the object's truthiness right before calling `logHmr`. this prevents
us from updating several accessor calls on `hmr`, the instance of the
`HotModuleReplacement` object, or wrapping the majority of the function
in an `if` statement.