chore(renovate): group download & upload artifact (#5193)
group github's `actions/download-artifact` and `actions/upload-artifact`
into a single renovate job. in #5191 and #5192, renovate created a pr
for each of these individually for the v4.0 release of each library.
however, they have a peer dependency on one another, making it such that
we could merge neither.

this was solved by cherry-picking the commit from the latter onto the
former. going forward, let's group these for all dependencies. we should
(hopefully) only run into this across majors, but we'll err on the
'safe' side of things for now