chore(deps): upgrade rollup/plugin-replace to v5
upgrade stencil's dependency on `@rollup/plugin-replace` from v2.x to
v5.x. this package is being done as a part of a larger effort to upgrade
stencil's version of rollup to v2 (which itself is a part of a larger
effort to move to v4). this package has been updated on its own, as:

1. doing so allows us to cut scope/size of other upgrade PRs
2. this package relies on rollup v1.20.0+, and no other sibling
   packages, making this upgrade less involved than others.