update replace plugin calls to use `preventAssignment`. `preventAssignment` was added in rollup v2.40.0. when we attempt to build stencil with the latest v5.x release of the replace-plugin, we received the following error:
```
(!) Plugin replace: @rollup/plugin-replace: 'preventAssignment'
currently defaults to false. It is recommended to set this option to
`true`, as the next major version will default this option to `true`.
```

explicitly set it to suppress this error. doing so does not change the
stencil binary itself, nor does it appear to change the output of the
generated code.