chore(deps): upgrade rollup/plugin-replace to v5 (#5241)
* chore(deps): upgrade rollup/plugin-replace to v5

upgrade stencil's dependency on `@rollup/plugin-replace` from v2.x to
v5.x. this package is being done as a part of a larger effort to upgrade
stencil's version of rollup to v2 (which itself is a part of a larger
effort to move to v4). this package has been updated on its own, as:

1. doing so allows us to cut scope/size of other upgrade PRs
2. this package relies on rollup v1.20.0+, and no other sibling
   packages, making this upgrade less involved than others.

* update replace plugin calls to use `preventAssignment`.
`preventAssignment` was added in rollup v2.40.0. when we attempt to
build stencil with the latest v5.x release of the replace-plugin, we
received the following error:

```
(!) Plugin replace: @rollup/plugin-replace: 'preventAssignment'
currently defaults to false. It is recommended to set this option to
`true`, as the next major version will default this option to `true`.
```

explicitly set it to suppress this error. doing so does not change the
stencil binary itself, nor does it appear to change the output of the
generated code.

* remove `@rollup/plugin-replace` & `rollup-plugin-sourcemaps` from renovate config
remove the former so that renovate can begin to update this package. do
the same for `rollup-plugin-sourcemaps`, which is already up to date.

STENCIL-595: Update Rollup v2.x Infrastructure