chore(repo): prevent renovate from upgrating open (#5218)
Open v10 has a single breaking change, in that it requires Node v18+ to
be used. Stencil v4 still supports Node 16, and must continue to use
Open v9 as a result. We cannot drop support for Node 16 without a
breaking change, so we’ll hold off on this work until Stencil v5.