feat(compiler): deprecate customResolveOptions config option (#5269)
deprecate the `nodeResolve#customResolveOptions` configuration option
in a Stencil configuration. this field is used to configure
`@rollup/plugin-node-resolve` when generating any non-documentation
output target artifacts. the contents of this option get propagated by
the rollup plugin to `resolve`, which does the actual resolution of
files in stencil's in-memory filesystem.

in newer versions of `@rollup/plugin-node-resolve`, this option has been
removed, which no option to override defaults (creating a separation of
concerns between the plugin and `resolve`). since stencil exposes this
configuration in its public api, we cannot remove it without causing a
breaking change (nor could we find a suitable way to reuse an existing
configuration).

this change is a prerequisite to upgrading `@rollup/plugin-node-resolve`.
we will not remove the configuration option/upgrade the plugin until
stencil v5. in the interim, we add a warning message at config
validation time to try to elicit feedback on this change

STENCIL-595: Update Rollup v2.X Infrastructure