fix(hmr): allow changes to component decorators when using HMR
This fixes an issue where HMR updates which changed arguments to the
`@Component` decorator would crash the application because a `HostRef`
for the component could not be located.

This occurred because HMR updates caused a new version of the
component's lazy bundle to be fetched, which in turn would bundle the
`getHostRef` function and the corresponding `hostRefs` variable, which
is declared here:

https://github.com/ionic-team/stencil/blob/0041dc27b9deb45cfd1434a3d82d563793e56056/src/client/client-host-ref.ts#L18

Because the newly fetched lazy entry point for the updated component did
not have access to the `hostRefs` variable which was in-scope in the
_old_ bundle, the `getHostRef` function would always return `undefined`,
on the first render after the HMR update, causing issues in functions
like `setValue` which make use of a `HostRef`.

In order to fix the issue we can persist the `hostRefs` map across HMR
updates by storing a reference to it on the `window` object. We can
additionally constrain this persistence to only occur when we're
building with HMR, so that we don't clutter up the `window` object of
unsuspecting Stencil users in the context of a production build.

STENCIL-973