chore(ci): pack a .tgz with the build and post a comment
This adds a new CI workflow which will download the build artifact for a
PR build, pack a new `.tgz` file containing the build artifacts using
`npm pack`, and then write a comment on the PR, providing a quick and
user-friendly way to download and then install a build of the PR in a
Stencil project.

The idea here is to make it easier for team members to test out a PR
without having to clone and branch and build locally.