chore(docs): add a step for creating a new version of the stencil docs (#5295)
This adds a step to the release flow which consists of creating a new
version of the docs on the stencil-site repo, if necessary. Since we
already have good documentation of how that process works over there
(https://github.com/ionic-team/stencil-site/blob/main/RELEASE.md#creating-a-new-version-section)
we can just point to that.