chore(config): add some documentation to the `buildDist` option (#5296)
This option will control effectively whether or not Stencil will process
all output targets or not. By default, it will be set to the opposite of
`devMode`, so if you build a Stencil project with `--dev` then
`buildDist` will be false, while for a prod build it will be true.
Previously this behavior wasn't documented anywhere.