fix(runtime): only generate lazy build CSS when there are component tags (#5305)
* fix(runtime): only generate lazy build CSS when there are component tags

This commit fixes an issue where invalid CSS could be generated and inserted into the DOM if `bootstrapLazy` was called with no component metadata or if no new components are registered in the custom element registry (i.e. it is called multiple times with the same metadata).

STENCIL-632

Closes: #3771

Co-authored-by: Robin Aldenhoven <raldenhoven@users.noreply.github.com>

* remove prehydration tests

This commit removes some "invisible prehydration" tests that no longer work based on changes in 2559917a62b5a6b99dbd3d70ac15876398336a55. They no longer work because `bootstrapLazy` is called with the same compiler metadata so styles are not re-inserted after being manually removed by calls to `tearDownStylesScripts` in the respective test files

* chore(karma): remove additional invisilbePrehydration infra

* mark todo

---------

Co-authored-by: Robin Aldenhoven <raldenhoven@users.noreply.github.com>
Co-authored-by: Ryan Waskiewicz <ryanwaskiewicz@gmail.com>