fix(mock-doc): improve error message when `:scope` selector is used (#5318)
This improves the error message shown when there is an error on the
jQuery side when running `querySelector` and friends by selectively
making an addendum to the error message to just explain a bit better
what's going on.

In particular, we know that the `:scope`, `:where`, and `:is`
pseudo-selectors aren't supported at present in jQuery, so we can detect
if the selector passed to our mock-doc impls of `querySelector` and co
included any of those selectors and, if so, add a message with some
specific info about what's going on and what to do.

STENCIL-1108