chore(ci): refactor pack-and-comment workflow to support community PRs
This changes the workflow from being a `workflow_call` triggered
workflow which is intended to be called from _another_ workflow (`main`)
and instead makes it a `pull_request_target` workflow which is
"standalone". This means that no special privileges will be necessary to
run it and that will in turn allow the `create-or-update-comment` action
to work properly.

See these resources for more details:

- https://github.com/peter-evans/create-pull-request/blob/main/docs/concepts-guidelines.md#restrictions-on-repository-forks
- https://github.blog/2020-08-03-github-actions-improvements-for-fork-and-pull-request-workflows/