chore(repo): remove old release scripts pt2
now that stencil releases are largely automated, release tasks that
were skipped during manual releases can be removed. this commit removes
all tasks that were skipped with `isCI` is was `true` (as it's always
`true` these days).

additional conditionals that use logical-AND with `isCI` have been
updated, as that part of the conditional is now always `true`.

remove the `--otp` flag, as it is only used for manual releases

as a result of these removals, the `--any-branch` flag can be safely
removed from our release pipeline, as its no longer used.