chore(build): make esbuild standalone (#5385)
This removes the dependence of the esbuild-based build on the existing
Rollup-based build. Previously we needed to run the Rollup-based build
first so that certain files which could not yet be built with Esbuild
would be present on disk, but since we can now build everything with
Esbuild we don't need to take this step anymore.

We do, however, continue to need to run the TypeScript compiler in order
to generate up-to-date typedef files.

Part of STENCIL-1016