feat(dev-server): make "ping" route configurable
This commit adds an option to the dev server config object as a part of the Stencil config to allow users to change the route for the "ping" response.

The user supplied route is validated when validating the dev server config. If the user sets the ping route to `null`, the route will not be registered.