chore(build): add `--prod` and `--ci` arguments to `build.esbuild`
This adds the `--prod` and `--ci` arguments to `build.esbuild` when the
Esbuild build script (source at `scripts/esbuild/build.ts`) is invoked.
This ensures that we're setting the `isProd` and `isCi` options in the
same way for the `build` and `build.esbuild` scripts in `package.json`.

also part of STENCIL-1016